waitless/api/requirements.txt
pymongo==4.5.0
dnspython==2.3.0
python-dotenv==1.0.0